Get relocated go_modules file updater working
